Refactored tests to remove obsolete project
