--- DFC.App.JobProfiles.HowToBecome.MessageFunctionAppTests/CustomWebApplicationFactory.cs
@@ -1,23 +1,0 @@
-﻿using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace DFC.App.JobProfiles.HowToBecome.MessageFunctionApp.IntegrationTests
-{
-    public class CustomWebApplicationFactory<TStartup> : WebApplicationFactory<TStartup>
-        where TStartup : class
-    {
-        protected override void ConfigureWebHost(IWebHostBuilder builder)
-        {
-            builder?.ConfigureServices(services =>
-            {
-                var configuration = new ConfigurationBuilder()
-                    .AddJsonFile("appsettings.json", optional: true, reloadOnChange: true)
-                    .Build();
-
-                services.AddSingleton<IConfiguration>(configuration);
-            });
-        }
-    }
-}
