--- DFC.App.JobProfiles.HowToBecome.MessageFunctionAppTests/appsettings-template.json
@@ -1,17 +1,0 @@
-{
-  "Configuration": {
-    "CosmosDbConnections": {
-      "JobProfileSegment": {
-        "AccessKey": "C2y6yDjf5/R+ob0N8A7Cgv30VRDJIWEHLM+4QDU5DE2nQ9nDuVTqobD4b8mGGyPMbIZnqyMsEcaGQy67XIw/Jw==",
-        "EndpointUrl": "https://localhost:8081",
-        "DatabaseId": "jobProfile_IntegrationTest",
-        "CollectionId": "howToBecomeSegments_IntegrationTest",
-        "PartitionKey": "/PartitionKey"
-      }
-    }
-  },
-  "ServiceBusOptions": {
-    "ServiceBusConnectionString": "[service bus connection string]",
-    "topicName": "job-profile-refresh"
-  }
-}
