+++ DFC.App.JobProfiles.HowToBecome.sln
@@ -47,19 +47,15 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DFC.App.JobProfiles.HowToBe
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ViewsTests", "ViewsTests", "{E84C1DE2-6776-4242-A228-0BEA41BEB2AB}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DFC.App.JobProfiles.HowToBecome.Views.UnitTests", "DFC.App.JobProfiles.HowToBecome.Views.Tests\DFC.App.JobProfiles.HowToBecome.Views.UnitTests.csproj", "{0F8BE735-7DC2-488B-A882-3CE599452484}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Functions", "Functions", "{536FC573-B3B1-4CE7-AF25-57C472E7F050}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DFC.App.JobProfiles.HowToBecome.MessageFunctionApp", "DFC.App.JobProfiles.HowToBecome.MessageFunctionApp\DFC.App.JobProfiles.HowToBecome.MessageFunctionApp.csproj", "{3E578F7F-787C-47D4-ACE9-8D3DD6505B00}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "MessagingFunction", "MessagingFunction", "{DA1540EF-9854-4211-9ACD-55F7F72B0174}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DFC.App.JobProfiles.HowToBecome.MessageFunctionApp.IntegrationTests", "DFC.App.JobProfiles.HowToBecome.MessageFunctionAppTests\DFC.App.JobProfiles.HowToBecome.MessageFunctionApp.IntegrationTests.csproj", "{BFEF351F-6772-4700-9CEB-620C02557656}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Functions", "Functions", "{7E42A252-88C9-484A-8CF8-035E59593201}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DFC.App.JobProfiles.HowToBecome.MFA.UnitTests", "DFC.App.JobProfiles.HowToBecome.MFA.UnitTests\DFC.App.JobProfiles.HowToBecome.MFA.UnitTests.csproj", "{7F971278-DDC5-4B04-B00B-31FA50456E13}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DFC.App.JobProfiles.HowToBecome.Views.UnitTests", "DFC.App.JobProfiles.HowToBecome.Views.UnitTests\DFC.App.JobProfiles.HowToBecome.Views.UnitTests.csproj", "{DE6EA966-3884-493A-B6D8-024C894D9295}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -95,22 +91,18 @@ Global
 		{1F22A12D-AE61-4A72-9CC6-F04AC39AF056}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1F22A12D-AE61-4A72-9CC6-F04AC39AF056}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1F22A12D-AE61-4A72-9CC6-F04AC39AF056}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0F8BE735-7DC2-488B-A882-3CE599452484}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0F8BE735-7DC2-488B-A882-3CE599452484}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0F8BE735-7DC2-488B-A882-3CE599452484}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0F8BE735-7DC2-488B-A882-3CE599452484}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3E578F7F-787C-47D4-ACE9-8D3DD6505B00}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{3E578F7F-787C-47D4-ACE9-8D3DD6505B00}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3E578F7F-787C-47D4-ACE9-8D3DD6505B00}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3E578F7F-787C-47D4-ACE9-8D3DD6505B00}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BFEF351F-6772-4700-9CEB-620C02557656}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BFEF351F-6772-4700-9CEB-620C02557656}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BFEF351F-6772-4700-9CEB-620C02557656}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BFEF351F-6772-4700-9CEB-620C02557656}.Release|Any CPU.Build.0 = Release|Any CPU
 		{7F971278-DDC5-4B04-B00B-31FA50456E13}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7F971278-DDC5-4B04-B00B-31FA50456E13}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7F971278-DDC5-4B04-B00B-31FA50456E13}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7F971278-DDC5-4B04-B00B-31FA50456E13}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DE6EA966-3884-493A-B6D8-024C894D9295}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DE6EA966-3884-493A-B6D8-024C894D9295}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DE6EA966-3884-493A-B6D8-024C894D9295}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DE6EA966-3884-493A-B6D8-024C894D9295}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -128,12 +120,10 @@ Global
 		{D3689FD2-C33B-4FCD-8E7B-AD3BD583D7FD} = {72A8418B-D692-4B06-8948-36F15756B35E}
 		{1F22A12D-AE61-4A72-9CC6-F04AC39AF056} = {D3689FD2-C33B-4FCD-8E7B-AD3BD583D7FD}
 		{E84C1DE2-6776-4242-A228-0BEA41BEB2AB} = {C1FB0E35-9A76-4AD5-82E2-706A040A3575}
-		{0F8BE735-7DC2-488B-A882-3CE599452484} = {E84C1DE2-6776-4242-A228-0BEA41BEB2AB}
 		{3E578F7F-787C-47D4-ACE9-8D3DD6505B00} = {536FC573-B3B1-4CE7-AF25-57C472E7F050}
-		{DA1540EF-9854-4211-9ACD-55F7F72B0174} = {72A8418B-D692-4B06-8948-36F15756B35E}
-		{BFEF351F-6772-4700-9CEB-620C02557656} = {DA1540EF-9854-4211-9ACD-55F7F72B0174}
 		{7E42A252-88C9-484A-8CF8-035E59593201} = {91D6C924-05D6-4F74-AEAC-1F3695B81ADD}
 		{7F971278-DDC5-4B04-B00B-31FA50456E13} = {7E42A252-88C9-484A-8CF8-035E59593201}
+		{DE6EA966-3884-493A-B6D8-024C894D9295} = {E84C1DE2-6776-4242-A228-0BEA41BEB2AB}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {696A5DAA-2A35-46E3-BBCD-C9B55866F0A7}
